Check in some merge of libspin with 3.1 release, not beta
git-svn-id: http://svn.open64.net/svnroot/open64@1323 ead328b1-e903-0410-a19d-eacee4b031f6

+++ branches/merge/libspin/gspin-mempool.h
@@ -47,6 +47,6 @@ extern gs_arena_t gs_mempool[GS_ARENA_COUNT];
 
 extern void *__gs_mempool_alloc(unsigned int mempool_id, unsigned int count);
 extern void *gs_mempool_idx2address(int arena_id, int cell_index);
-extern long gs_mempool_address2byteofst(int arena_id, char *p);
+extern int gs_mempool_address2byteofst(int arena_id, char *p);
 
 #endif // __GSPIN_GS_MEMPOOL_H__
